% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaGrid.R
\name{create_usa_grid}
\alias{create_usa_grid}
\title{Create a grid of points that fall within the contiguous USA.}
\usage{
create_usa_grid(resolution = 1)
}
\arguments{
\item{resolution}{optional argument for controlling the resolution, defaults to 1}
}
\value{
A spatial dataframe containing the grid points within the contiguous USA.
}
\description{
Create a grid of points that fall within the contiguous USA.
}
\examples{

result <- create_usa_grid()
print(result)
plot(result)
dev.off()
result05 <- create_usa_grid(resolution = 0.5)
print(result05)
plot(result05)
}
