% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{get_temp_trend}
\alias{get_temp_trend}
\title{Estimating yearly trend of temperatures over time}
\usage{
get_temp_trend(data, station)
}
\arguments{
\item{data}{data frame with NCEI station data}

\item{station}{name of station}
}
\value{
a data frame with following components:
# example code
}
\description{
Retrieve yearly average temperature for a given station, in degrees Celsius
}
