% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyCycle.R
\name{get_yearly_cycle}
\alias{get_yearly_cycle}
\title{Calculate and return the average daily temperatures for each day of the year for a specified weather station.}
\usage{
get_yearly_cycle(data, station_name)
}
\arguments{
\item{data}{data frame with NCEI data}

\item{station}{station name}
}
\value{
A data frame
}
\description{
Calculate and return the average daily temperatures for each day of the year for a specified weather station.
}
