% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_station_to_grid}
\alias{interpolate_station_to_grid}
\title{Interpolate station data to a grid within the contiguous USA using different methods}
\usage{
interpolate_station_to_grid(data_stations, grid, var)
}
\arguments{
\item{data_stations}{A dataframe containing station locations and the variable to interpolate.}

\item{grid}{A spatial dataframe of grid points where data needs to be interpolated.}

\item{var}{The name of the variable in data_stations to interpolate.}

\item{method}{The regression method used, defaults to Gaussian process (gp)}
}
\value{
A spatial dataframe of grid points with interpolated values.
}
\description{
Interpolate station data to a grid within the contiguous USA using different methods
}
\examples{
usa_grid <- create_usa_grid(resolution = 5)
interpolated_data <- interpolate_station_to_grid(data, points, "T_DAILY_AVG")
print(interpolated_data)
}
