% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_to_grid}
\alias{interpolate_to_grid}
\title{Interpolate station data to a grid within the contiguous USA}
\usage{
interpolate_to_grid(station_data, variable_name, grid_points, idw_power = 2)
}
\arguments{
\item{station_data}{A spatial dataframe containing station locations and the variable to interpolate.}

\item{variable_name}{The name of the variable in station_data to interpolate.}

\item{grid_points}{A spatial dataframe of grid points where data needs to be interpolated.}

\item{idw_power}{The power parameter for IDW interpolation, default is 2.}
}
\value{
A spatial dataframe of grid points with interpolated values.
}
\description{
Interpolate station data to a grid within the contiguous USA
}
\examples{
stations <- st_as_sf(data.frame(longitude = c(-120, -122), latitude = c(36, 38), temp = c(15, 10)),
                     coords = c("longitude", "latitude"), crs = 4326)
grid <- create_usa_grid(resolution = 1)
interpolated_data <- interpolate_to_grid(stations, "temp", grid)
print(interpolated_data)
}
