% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_to_grid}
\alias{interpolate_to_grid}
\title{Interpolate station data to a grid within the contiguous USA}
\usage{
interpolate_to_grid(data, data_to_predict, variable)
}
\arguments{
\item{variable}{The target variable}

\item{station_data}{A spatial dataframe containing station locations and the variable to interpolate.}

\item{date_to_predict}{The date to predict}
}
\value{
A spatial dataframe of grid points with interpolated values.
}
\description{
Interpolate station data to a grid within the contiguous USA
}
\examples{
stations <- st_as_sf(data.frame(longitude = c(-120, -122), latitude = c(36, 38), temp = c(15, 10)),
                     coords = c("longitude", "latitude"), crs = 4326)
grid <- create_usa_grid(resolution = 1)
interpolated_data <- interpolate_to_grid(stations, "temp", grid)
print(interpolated_data)
}
